feat: allow app theme mixins
This may not be necessary for anyone using variables directly or the convenience mixins provided by this library, but it will be helpful for anyone who is still using traditional scss mixins to do theming in their app. It is also the best way for an app theme to access $is-dark or items in the background/foreground palettes.